Extract RestRequest and URIBuilder to a network module
