+++ bot/build.gradle
@@ -31,6 +31,7 @@ module {
 
 dependencies {
     implementation project(':host')
+    implementation project(':network')
     implementation project(':vcs')
     implementation project(':json')
     implementation project(':census')
