+++ bot/src/main/java/module-info.java
@@ -24,6 +24,7 @@ module org.openjdk.skara.bot {
     requires transitive org.openjdk.skara.host;
     requires transitive org.openjdk.skara.json;
     requires transitive org.openjdk.skara.census;
+    requires org.openjdk.skara.network;
     requires org.openjdk.skara.vcs;
     requires java.logging;
     requires jdk.httpserver;
