+++ bots/hgbridge/src/main/java/module-info.java
@@ -23,6 +23,7 @@
 module org.openjdk.skara.bots.hgbridge {
     requires org.openjdk.skara.vcs;
     requires org.openjdk.skara.host;
+    requires org.openjdk.skara.network;
     requires org.openjdk.skara.census;
     requires org.openjdk.skara.bot;
     requires org.openjdk.skara.process;
