+++ bots/hgbridge/src/main/java/org/openjdk/skara/bots/hgbridge/JBridgeBotFactory.java
@@ -23,7 +23,7 @@
 package org.openjdk.skara.bots.hgbridge;
 
 import org.openjdk.skara.bot.*;
-import org.openjdk.skara.host.network.URIBuilder;
+import org.openjdk.skara.network.URIBuilder;
 import org.openjdk.skara.json.*;
 
 import java.util.*;
