+++ bots/notify/src/main/java/module-info.java
@@ -27,6 +27,7 @@ module org.openjdk.skara.bots.notify {
     requires org.openjdk.skara.email;
     requires org.openjdk.skara.storage;
     requires org.openjdk.skara.mailinglist;
+    requires org.openjdk.skara.network;
     requires java.logging;
 
     provides org.openjdk.skara.bot.BotFactory with org.openjdk.skara.bots.notify.JNotifyBotFactory;
