+++ host/build.gradle
@@ -40,6 +40,7 @@ dependencies {
     implementation project(':ini')
     implementation project(':process')
     implementation project(':email')
+    implementation project(':network')
 
     testImplementation project(':test')
 }
