+++ host/src/main/java/org/openjdk/skara/host/github/GitHubApplication.java
@@ -22,7 +22,7 @@
  */
 package org.openjdk.skara.host.github;
 
-import org.openjdk.skara.host.network.URIBuilder;
+import org.openjdk.skara.network.URIBuilder;
 import org.openjdk.skara.json.*;
 
 import java.io.*;
