+++ host/src/test/java/org/openjdk/skara/host/github/GitHubHostTests.java
@@ -22,7 +22,7 @@
  */
 package org.openjdk.skara.host.github;
 
-import org.openjdk.skara.host.network.URIBuilder;
+import org.openjdk.skara.network.URIBuilder;
 import org.openjdk.skara.test.TemporaryDirectory;
 
 import org.junit.jupiter.api.Test;
