+++ network/src/main/java/org/openjdk/skara/network/URIBuilder.java
@@ -20,7 +20,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package org.openjdk.skara.host.network;
+package org.openjdk.skara.network;
 
 import java.io.UnsupportedEncodingException;
 import java.net.*;
