+++ network/src/test/java/org/openjdk/skara/network/RestRequestTests.java
@@ -20,10 +20,11 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package org.openjdk.skara.host.network;
+package org.openjdk.skara.network;
 
 import com.sun.net.httpserver.*;
 import org.openjdk.skara.json.*;
+import org.openjdk.skara.network.*;
 
 import org.junit.jupiter.api.Test;
