+++ settings.gradle
@@ -40,6 +40,7 @@ include 'ssh'
 include 'test'
 include 'vcs'
 include 'webrev'
+include 'network'
 
 include 'bots:cli'
 include 'bots:forward'
