+++ storage/src/main/java/module-info.java
@@ -21,6 +21,7 @@
  * questions.
  */
 module org.openjdk.skara.storage {
+    requires org.openjdk.skara.network;
     requires org.openjdk.skara.host;
     requires org.openjdk.skara.vcs;
