+++ test/src/main/java/module-info.java
@@ -29,6 +29,7 @@ module org.openjdk.skara.test {
     requires org.openjdk.skara.bot;
     requires org.openjdk.skara.json;
     requires org.openjdk.skara.host;
+    requires org.openjdk.skara.network;
     requires org.openjdk.skara.email;
     requires org.openjdk.skara.mailinglist;
     requires org.openjdk.skara.proxy;
